+++ emukit/core/interfaces/models.py
@@ -46,25 +46,51 @@ class IDifferentiable:
         Computes and returns model gradients of mean and variance at given points
 
         :param X: points to compute gradients at
+        :returns: Tuple of gradients of mean and variance.
+        """
+        raise NotImplementedError
+
+
+class IJointlyDifferentiable:
+    def predict_with_full_covariance(self, X: np.ndarray) -> Tuple[np.ndarray, np.ndarray]:
+        """
+        :param X: (n_points x n_dimensions) array containing locations at which to get predictions
+        :return: (mean, variance) Arrays of size n_points x 1 and n_points x n_points of the predictive
+                 mean and variance at each input location
+        """
+        raise NotImplementedError
+
+    def get_joint_prediction_gradients(self, X: np.ndarray) -> Tuple[np.ndarray, np.ndarray]:
+        """
+        Computes and returns model gradients of mean and full covariance matrix at given points
+
+        :param X: points to compute gradients at, nd array of shape (q, d)
+        :return: Tuple with first item being gradient of the mean of shape (q) at X with respect to X (return shape is (q, q, d)).
+                 The second item is the gradient of the full covariance matrix of shape (q, q) at X with respect to X
+                 (return shape is (q, q, q, d)).
         """
         raise NotImplementedError
 
 
 class IPriorHyperparameters:
-    def generate_hyperparameters_samples(self, n_samples: np.int) -> np.ndarray:
+    def generate_hyperparameters_samples(self, n_samples: int, n_burnin: int,
+                                         subsample_interval: int, step_size: float, leapfrog_steps: int) -> np.ndarray:
         """
         Generates the samples from the hyper-parameters of the model.
 
+        :param n_samples: Number of hyper-parameter samples
+        :param n_burnin: Number of initial samples not used.
+        :param subsample_interval: Interval of subsampling from HMC samples.
+        :param step_size: Size of the gradient steps in the HMC sampler.
+        :param leapfrog_steps: Number of gradient steps before each Metropolis Hasting step.
         :return: numpy array whose rows are samples from the hyper-parameters of the model.
         """
         raise NotImplementedError
 
-    def fix_model_hyperparameters(self,sample_hyperparameters: np.ndarray) -> None:
+    def fix_model_hyperparameters(self, sample_hyperparameters: np.ndarray) -> None:
         """
         Fixes the model hyper-parameters to certain values (which can be taken from samples).
 
         :param sample_hyperparameters: np.ndarray whose rows contain each hyper-parameters set.
         """
         raise NotImplementedError
-
-