+++ tests/emukit/test_acquisitions.py
@@ -7,6 +7,7 @@ from scipy.optimize import check_grad
 
 from bayesian_optimization.test_entropy_search import entropy_search_acquisition
 from emukit.bayesian_optimization.acquisitions import ExpectedImprovement, NegativeLowerConfidenceBound, EntropySearch
+from emukit.bayesian_optimization.acquisitions import MaxValueEntropySearch
 from emukit.core.acquisition import IntegratedHyperParameterAcquisition
 from emukit.bayesian_optimization.acquisitions.entropy_search import MultiInformationSourceEntropySearch
 from emukit.bayesian_optimization.acquisitions.log_acquisition import LogAcquisition
@@ -15,9 +16,9 @@ from emukit.core.acquisition.acquisition_per_cost import CostAcquisition
 
 from emukit.bayesian_optimization.acquisitions import ProbabilityOfImprovement
 from emukit.bayesian_optimization.acquisitions import ProbabilityOfFeasibility
-from emukit.experimental_design.model_based.acquisitions import ModelVariance, IntegratedVarianceReduction
-from emukit.model_wrappers.gpy_quadrature_wrappers import convert_gpy_model_to_emukit_model
-from emukit.quadrature.acquisitions import SquaredCorrelation
+from emukit.experimental_design.acquisitions import ModelVariance, IntegratedVarianceReduction
+from emukit.model_wrappers.gpy_quadrature_wrappers import create_emukit_model_from_gpy_model
+from emukit.quadrature.acquisitions import SquaredCorrelation, MutualInformation, UncertaintySampling
 from emukit.quadrature.methods import VanillaBayesianQuadrature
 
 
@@ -34,7 +35,10 @@ acquisition_tests = [acquisition_test_tuple('negative_lower_confidence_bound_acq
                      acquisition_test_tuple('probability_of_improvement_acquisition', True, default_grad_tol),
                      acquisition_test_tuple('model_variance_acquisition', True, 1e-5),
                      acquisition_test_tuple('squared_correlation_acquisition', True, 1e-3),
+                     acquisition_test_tuple('mutual_information_acquisition', True, 1e-3),
+                     acquisition_test_tuple('uncertainty_sampling_acquisition', True, 1e-3),
                      acquisition_test_tuple('entropy_search_acquisition', False, np.nan),
+                     acquisition_test_tuple('max_value_entropy_search_acquisition', False, np.nan),
                      acquisition_test_tuple('multi_source_entropy_search_acquisition', False, np.nan),
                      acquisition_test_tuple('integrated_variance_acquisition', False, np.nan),
                      acquisition_test_tuple('integrated_expected_improvement_acquisition', True, default_grad_tol),
@@ -42,12 +46,12 @@ acquisition_tests = [acquisition_test_tuple('negative_lower_confidence_bound_acq
                      acquisition_test_tuple('probability_of_feasibility_acquisition', True, default_grad_tol)]
 
 
-# Vanilla bq model for squared correlation test
+# Vanilla bq model used to test bq acquisition functions
 @pytest.fixture
 def vanilla_bq_model(gpy_model, continuous_space, n_dims):
     integral_bounds = continuous_space.get_bounds()
-    model = convert_gpy_model_to_emukit_model(gpy_model.model, integral_bounds)
-    return VanillaBayesianQuadrature(model)
+    model = create_emukit_model_from_gpy_model(gpy_model.model, integral_bounds, None)
+    return VanillaBayesianQuadrature(model, model.X, model.Y)
 
 
 # Acquisition function fixtures
@@ -70,12 +74,12 @@ def integrated_expected_improvement_acquisition(gpy_model_mcmc):
 def integrated_probability_of_improvement_acquisition(gpy_model_mcmc):
     return IntegratedHyperParameterAcquisition(gpy_model_mcmc, ProbabilityOfImprovement, 10)
 
-  
+
 @pytest.fixture
 def probability_of_feasibility_acquisition(gpy_model):
     return ProbabilityOfFeasibility(gpy_model)
 
-  
+
 @pytest.fixture
 def cost_acquisition(gpy_model):
     return CostAcquisition(gpy_model, 1e-6)
@@ -92,6 +96,11 @@ def probability_of_improvement_acquisition(gpy_model):
 
 
 @pytest.fixture
+def max_value_entropy_search_acquisition(gpy_model, continuous_space):
+    return MaxValueEntropySearch(gpy_model, continuous_space, num_samples=2, grid_size=100)
+
+
+@pytest.fixture
 def model_variance_acquisition(gpy_model):
     return ModelVariance(gpy_model)
 
@@ -104,6 +113,17 @@ def integrated_variance_acquisition(gpy_model, continuous_space):
 @pytest.fixture
 def squared_correlation_acquisition(vanilla_bq_model):
     return SquaredCorrelation(vanilla_bq_model)
+
+
+@pytest.fixture
+def mutual_information_acquisition(vanilla_bq_model):
+    return MutualInformation(vanilla_bq_model)
+
+
+@pytest.fixture
+def uncertainty_sampling_acquisition(vanilla_bq_model):
+    return UncertaintySampling(vanilla_bq_model)
+
 
 @pytest.fixture
 @pytest.mark.parametrize('n_dims', [2])